# proto-2-springboot/config-delegate/src/main/java/com/example/users/generated/model/GoogleProtobufAny.java
package com.example.users.generated.model;

import java.net.URI;
import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import org.springframework.lang.Nullable;
import org.openapitools.jackson.nullable.JsonNullable;
import java.time.OffsetDateTime;
import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import io.swagger.v3.oas.annotations.media.Schema;


import java.util.*;
import jakarta.annotation.Generated;

import java.util.Map;
import java.util.HashMap;
import com.fasterxml.jackson.annotation.JsonAnyGetter;
import com.fasterxml.jackson.annotation.JsonAnySetter;
/**
 * Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
 */

@Schema(name = "GoogleProtobufAny", description = "Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.")
@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-03-16T20:49:57.455763847-05:00[America/Chicago]", comments = "Generator version: 7.12.0")
public class GoogleProtobufAny {

  private @Nullable String atType;

  public GoogleProtobufAny atType(String atType) {
    this.atType = atType;
    return this;
  }

  /**
   * The type of the serialized message.
   * @return atType
   */
  
  @Schema(name = "@type", description = "The type of the serialized message.", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
  @JsonProperty("@type")
  public String getAtType() {
    return atType;
  }

  public void setAtType(String atType) {
    this.atType = atType;
  }
    /**
    * A container for additional, undeclared properties.
    * This is a holder for any undeclared properties as specified with
    * the 'additionalProperties' keyword in the OAS document.
    */
    private Map<String, Object> additionalProperties;

    /**
    * Set the additional (undeclared) property with the specified name and value.
    * If the property does not already exist, create it otherwise replace it.
    */
    @JsonAnySetter
    public GoogleProtobufAny putAdditionalProperty(String key, Object value) {
        if (this.additionalProperties == null) {
            this.additionalProperties = new HashMap<String, Object>();
        }
        this.additionalProperties.put(key, value);
        return this;
    }

    /**
    * Return the additional (undeclared) property.
    */
    @JsonAnyGetter
    public Map<String, Object> getAdditionalProperties() {
        return additionalProperties;
    }

    /**
    * Return the additional (undeclared) property with the specified name.
    */
    public Object getAdditionalProperty(String key) {
        if (this.additionalProperties == null) {
            return null;
        }
        return this.additionalProperties.get(key);
    }

  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    GoogleProtobufAny googleProtobufAny = (GoogleProtobufAny) o;
    return Objects.equals(this.atType, googleProtobufAny.atType) &&
    Objects.equals(this.additionalProperties, googleProtobufAny.additionalProperties);
  }

  @Override
  public int hashCode() {
    return Objects.hash(atType, additionalProperties);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class GoogleProtobufAny {\n");
    sb.append("    atType: ").append(toIndentedString(atType)).append("\n");
    
    sb.append("    additionalProperties: ").append(toIndentedString(additionalProperties)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}