# proto-2-springboot/config-delegate/target/classes/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: gnostic@google.com
    name: Contact Name
    url: https://github.com/google/gnostic
  description: Description from annotation
  license:
    name: Apache License
    url: https://github.com/google/gnostic/blob/master/LICENSE
  title: Title from annotation
  version: Version from annotation
servers:
- url: /
tags:
- name: UserSvc
paths:
  /usrsvc/getuser:
    post:
      operationId: UserSvc_GetUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
          description: Default error response
      tags:
      - UserSvc
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: UserSvc
components:
  schemas:
    GoogleProtobufAny:
      additionalProperties: true
      description: Contains an arbitrary serialized message along with a @type that
        describes the type of the serialized message.
      example:
        '@type': '@type'
      properties:
        '@type':
          description: The type of the serialized message.
          type: string
      type: object
    Status:
      description: "The `Status` type defines a logical error model that is suitable\
        \ for different programming environments, including REST APIs and RPC APIs.\
        \ It is used by [gRPC](https://github.com/grpc). Each `Status` message contains\
        \ three pieces of data: error code, error message, and error details. You\
        \ can find out more about this error model and how to work with it in the\
        \ [API Design Guide](https://cloud.google.com/apis/design/errors)."
      example:
        code: 6
        details:
        - '@type': '@type'
        - '@type': '@type'
        message: message
      properties:
        code:
          description: "The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code]."
          format: int32
          type: integer
        message:
          description: "A developer-facing error message, which should be in English.\
            \ Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details]\
            \ field, or localized by the client."
          type: string
        details:
          description: A list of messages that carry the error details.  There is
            a common set of message types for APIs to use.
          items:
            $ref: '#/components/schemas/GoogleProtobufAny'
          type: array
      type: object
    UserRequest:
      example:
        userId: 0
      properties:
        userId:
          format: int64
          type: integer
      type: object
    UserResponse:
      example:
        fname: fname
        lname: lname
        userId: 0
      properties:
        userId:
          format: int64
          type: integer
        fname:
          type: string
        lname:
          type: string
      type: object
  securitySchemes:
    BasicAuth:
      scheme: basic
      type: http